Finally fixing text stacking on itself
When you set PS1 to a command, it will execute. No problem. But, if you
print something to the screen in color without escaping the colors it
can cause text wrapping and stacking issues. “Characters printed
outside of PS1 are not counted by Bash, which will cause it to
incorrectly place the cursor and clear characters” (ArchWiki).

http://stackoverflow.com/questions/11831296/os-x-terminal-text-stacking-
on-top-of-itself

https://wiki.archlinux.org/index.php/Bash/Prompt_customization#Embedding
_commands